requirements.txt
flask
flask-sqlalchemy
requests

pymysql
Flask-Migrate
authlib
python-dotenv